+++ src/styles/theme.tsx
@@ -2,4 +2,5 @@ export const Theme = {
   gray: '#E7E7E7',
   primary: '#254463',
   secondary: '#DFEFFF',
+  tertiary: '#95B4D3',
 };