+++ docs/changelog.md
@@ -35,6 +35,8 @@ Sprint 11 / April 26, 2020
 - removed error log from failed login attempt on Login page.
 - updated Graph component to adjust inner bar divisions based on records.
 - moved DashboardContext further out so that Graph component will be within its scope.
+- added tertiary theme color.
+- added logo and subheading to Login page.
 
 <hr />